* commit 'd95ecf36fda1c81bd76ddd260c7b71b8ac1402af': (76 commits)
  Corrected OfferVersion property from plural to singular
  Updated Dtos due to DBChange
  Updates to contract.
  Data for versions.
  Updated a typo on the the dto file
  equipment category added to equipmentmodel
  added equipment options to equipmentffertrndto
  Updated Dtos due to the changes in DB
  update for contract length description on list of offers
  Made some changes on Equipment related Dtos to save the offer
  Updated the Dtos
  Updated OfferDto and AcquiringOfferDto
  Dto Update
  Updated the Dtos
  .
  Updated the Dtos
  update session guid added to merchantofferDto
  Updated OfferDto
  Updated OfferDto
  Updated OfferDto
  ...

+++ PS.Mothership.Core/PS.Mothership.Core.Common/Dto/Merchant/ContactDto.cs
@@ -17,7 +17,7 @@ namespace PS.Mothership.Core.Common.Dto.Merchant
         public Guid ContactGuid { get; set; }
 
         [DataMember]
-        public long SalutationKey { get; set; }
+        public int SalutationKey { get; set; }
 
         [DataMember]
         public string FirstName { get; set; }
